+++ src/const.js
@@ -1,1 +1,17 @@
-export const FILM_COUNT_PER_STEP = 5;
+const FILM_COUNT_PER_STEP = 5;
+
+const FILTER_TYPE = {
+  ALL: 'All movies',
+  WATCHLIST: 'Watchlist',
+  HISTORY: 'History',
+  FAVORITES: 'Favorites',
+};
+
+const EMOTIONS = [
+  'smile',
+  'sleeping',
+  'puke',
+  'angry',
+];
+
+export {FILM_COUNT_PER_STEP, FILTER_TYPE, EMOTIONS};
