--- src/mock/filter.js
@@ -1,8 +1,0 @@
-import {filter} from '../utils/filter.js';
-
-export const generateFilter = (films) => Object.entries(filter).map(
-  ([filterName, filterFilms]) => ({
-    name: filterName,
-    count: filterFilms(films).length,
-  }),
-);
